Replace linear search with binary search
This gives a small speed improvement by using binary search. Most of the code
in this patch handles implementing halfodd() for different Lua versions, which
is used to calculate the next middle point for the binary search. The function
calculates the division of a number by two, and rounds it up to the next odd
number. This is because data tables have the following shape:

  tab = {
     lowbound1, upperbound1,
     lowbound2, upperbound2,
     ...
     lowboundN, upperboundN,
  }

The low bound of each rune range is always in an odd table index. The binary
search starts with the 1 as the index of the lower, and #tab as the upper
bound. The lower bound always contains odd indexes, and the upper bound always
contains even indexes.

The halfodd() function returns is used to calculate an odd middle index, and
then the rune being tested is compared with the values at tab[middle] (low
bound of the range) and tab[middle+1] (upper bound of the range). On each
iteration either the upper bound is updated to middle-1, or the lower bound to
middle+2, to maintain the invariant that lower bound indexes are always odd,
and upper bound indexes are always even.

Binary search increases the speed of wcwidth() from 0.566 milliseconds per
function call (average) to just 0.061 milliseconds. This is a 9.3x speedup.